Babel Bump & Prop Types & Lint

+++ src/containers/views/Documents.js
@@ -1,4 +1,5 @@
-import React, { Component, PropTypes } from 'react';
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 import { Link } from 'react-router';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
@@ -11,7 +12,7 @@ import Button from '../../components/Button';
 import {
   fetchCollection,
   deleteDocument,
-  filterBySearchInput
+  filterBySearchInput,
 } from '../../ducks/collections';
 import { search } from '../../ducks/utils';
 import { capitalize } from '../../utils/helpers';
@@ -38,9 +39,7 @@ export class Documents extends Component {
   handleClickDelete(filename) {
     const { deleteDocument, params } = this.props;
     const confirm = window.confirm(getDeleteMessage(filename));
-    if (confirm) {
-      deleteDocument(params.collection_name, params.splat, filename);
-    }
+    confirm && deleteDocument(params.collection_name, params.splat, filename);
   }
 
   renderTable() {
@@ -180,12 +179,12 @@ Documents.propTypes = {
   fetchCollection: PropTypes.func.isRequired,
   deleteDocument: PropTypes.func.isRequired,
   search: PropTypes.func.isRequired,
-  params: PropTypes.object.isRequired
+  params: PropTypes.object.isRequired,
 };
 
 const mapStateToProps = state => ({
   documents: filterBySearchInput(state.collections.entries, state.utils.input),
-  isFetching: state.collections.isFetching
+  isFetching: state.collections.isFetching,
 });
 
 const mapDispatchToProps = dispatch =>
@@ -193,7 +192,7 @@ const mapDispatchToProps = dispatch =>
     {
       fetchCollection,
       deleteDocument,
-      search
+      search,
     },
     dispatch
   );
